Refactor lightning code
